refactor: clean and move adapter and port packages for clarity
